+++ packages/react-components/.storybook/preview.js
@@ -1,0 +1,14 @@
+import { INITIAL_VIEWPORTS } from '@storybook/addon-viewport'
+
+export const parameters = {
+  actions: { argTypesRegex: '^on[A-Z].*' },
+  controls: {
+    matchers: {
+      color: /(background|color)$/i,
+      date: /Date$/,
+    },
+  },
+  viewport: {
+    viewports: INITIAL_VIEWPORTS,
+  },
+}