+++ packages/react-components/src/icon/stories/member.stories.js
@@ -1,0 +1,18 @@
+import React from 'react'
+import { Member } from '../index'
+import releaseBranch from '@twreporter/core/lib/constants/release-branch'
+const { master, staging, preview, release } = releaseBranch
+
+export default {
+  title: 'Icon/Member',
+  component: Member,
+  argTypes: {
+    releaseBranch: {
+      defaultValue: master,
+      options: [master, staging, preview, release],
+      control: { type: 'radio' },
+    },
+  },
+}
+
+export const member = args => <Member {...args} />