+++ packages/react-article-components/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.11-rc.2](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-article-components@1.4.11-rc.1...@twreporter/react-article-components@1.4.11-rc.2) (2022-05-09)
+
+**Note:** Version bump only for package @twreporter/react-article-components
+
+
+
+
+
+## [1.4.11-rc.1](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-article-components@1.4.11-rc.0...@twreporter/react-article-components@1.4.11-rc.1) (2022-05-09)
+
+**Note:** Version bump only for package @twreporter/react-article-components
+
+
+
+
+
+## [1.4.11-rc.0](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-article-components@1.4.10...@twreporter/react-article-components@1.4.11-rc.0) (2022-05-03)
+
+**Note:** Version bump only for package @twreporter/react-article-components
+
+
+
+
+
 ## [1.4.10](https://github.com/twreporter/twreporter-npm-packages/compare/@twreporter/react-article-components@1.4.10-rc.0...@twreporter/react-article-components@1.4.10) (2022-04-29)
 
 **Note:** Version bump only for package @twreporter/react-article-components